Rename library to fix name misspelling
