+++ main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"github.com/ncaak/roll-the-dices/lib/conn"
-	"github.com/ncaak/roll-the-dices/lib/dices"
+	"github.com/ncaak/roll-the-dices/lib/dice"
 	"github.com/ncaak/roll-the-dices/lib/storage"
 	"log"
 	"fmt"
@@ -18,7 +18,7 @@ func main() {
 	var offset = storage.GetUpdateOffset()
 	var messages = conn.GetUpdates(offset)
 
-log.Print(dices.New())
+log.Print(dice.New())
 	for _, msg := range messages {
 
 		if msg.IsCommand() == true {
@@ -30,13 +30,13 @@ log.Print(dices.New())
 				var argument = strings.TrimSpace(command[2])
 				switch command[1] {
 				case acceptedCommands[0]:
-					reply = dices.Roll(argument)
+					reply = dice.Roll(argument)
 
 				case acceptedCommands[1]:
-					reply = dices.Advantage(argument)
+					reply = dice.Advantage(argument)
 
 				case acceptedCommands[2]:
-					reply = dices.Disadvantage(argument)
+					reply = dice.Disadvantage(argument)
 				}
 
 				//conn.SendReply(msg.Message.Chat.Id, reply, msg.Message.MessageId)
